Simplify the storage sample.
A user complained it was too confusing, especially with all the
authentication code.  So, I reduced the code to show only one method of
auth.  This is a cloud storage sample, not an auth sample.

Also, I moved the // [START all] and // [END all] region tags to contain a
smaller window of code.  Now that the View on GitHub button is always
there, there is little value in showing boilerplate code.