Fix flaky tests with true deterministic clean up.
Tests were failing, even with a 2-second delay before calling list.  2
seconds is not long enough.  So, use Eventually() to give assertions
more than 10 seconds to become true, but not delay when the assertion
succeeds immedately.  And explicitly track every file that must be deleted
at the end of a test instead of listing them.