Migrate dotnet core project files to VS 2017.
Remove old *.xproj and project.json files.

When this change is committed, the jenkins build will break until I
update the build machines with the latest version of the dotnet sdk.
I haven't found a way to migrate the project files without breaking
something for some small period of time.