Move doc tags in Stackdriver sample to show whole Configuration functions.
The order in which these functions are called matters.  For example, app.UseGoogleTrace() must be called before UseGoogleTrace(), but the code snippet did not demonstrate that, because it only showed the one line of code app.UseGoogleTrace()).

I moved the doc tags so that it shows the whole  Configure() function.  This should give readers more context.