Combine all the IoT solutions and add all dependency projects
Previously each project had its own solution file, and although each project referenced other projects (so it could build from the command line) Visual Studio requires referenced projects to be part of the same solution. Using a single solution fixes this, and is also more convenient for users with a single obvious file to open.

This still gives a lot of warnings due to dependency *packages* that only target .NET Framework instead of .NET Core, but the solution does open and build.