Spanner: Use correct/consistent logic to transfer budget.
The `WriteWithTransactionUsingDmlCoreAsync` snippet moves $200,000 from Album1 to Album2. However, Album1 is initialized with a budget of $100,000. As a result, you can't work through the [getting started doc](https://cloud.google.com/spanner/docs/getting-started/csharp/) from start to finish. This snippet also defined and checked a `minimumAmountToTransfer`, which isn't consistent with the equivalent snippets for other languages.

In addition, the `ReadWriteWithTransactionCoreAsync` snippet moved $200,000, but in the opposite direction, which was somewhat confusing.

Finally, neither snippet confirmed that the source album's budget was at least as large as the transfer amount.

I made both snippets transfer $100,000 from Album1 to Album2. I also fixed the logic errors and updated the tests as needed.