Spanner: Prevent crash when an Album has a null MarketingBudget
If you're working through the "Getting started in C#" docs [1] from start to finish, then you'll see a crash when you run `dotnet run queryDataWithIndex`. That's because one of the albums in the query results has a null MarketingBudget, and QueryDataWithIndexAsync() tries to format the DBNull as a string. (I'm using .NET Core on a Mac; I don't know whether this happens with .NET Framework on Windows.)

This change checks for a null MarketingBudget and prints it as an empty string. It also adds a test for QueryDataWithIndexAsync().

[1]: https://cloud.google.com/spanner/docs/getting-started/csharp/