Update samples with startup class changes
I've left the Slack example using an assembly-attributed startup,
partly because it's better given the structure of the sample, and
partly so that we have at least *one* example of the
assembly-attributed startups.

The HttpClient change brings us into a more conventional approach to
obtaining HttpClients via dependency injection, I believe.