feat: add NUMERIC column to table with all data types
The spanner_create_table_with_datatypes should create a table with all
supported data types in Cloud Spanner. The .NET samples (as well as some
other client languages) missed a NUMERIC column as this data type was
added later than the originally supported data types. This change adds a
NUMERIC column to the sample to bring it up to date with the samples of
the other client languages, and to bring it in line with the description
of the sample.

Fixes #1254