test: retry CreateInstance test on Unavailable
The CreateInstance test should be retried if the error that is returned is Unavailable.
Unavailable can be returned by any RPC, and should normally be retried automatically for
idem-potent RPCs. CreateInstance is not idem-potent, and is therefore not retried
automatically. In this test however we can safely retry it without checking the status of
the backend, as the test expects an AlreadyExists error to be returned.

Fixes #1332