feat(appengine): Delete AntiForgery sample (#2204)
Context: I saw [this
blog](https://medium.com/google-cloud/antiforgery-tokens-asp-net-core-and-google-cloud-7ac6a5c7842b)
which provides a motivating example for the sample. I gave it a shot on
GAE Flex without doing any of that Storage or KMS stuff, but keeping the
`ValidateAntiForgeryToken` attribute and `AddDataProtection()`. It
_still_ worked. I saw the custom home page, entered some text, clicked
submit (http post) and got the response with no errors.

To me, this sample is ok to remove -- perhaps it was inspired by trying
to work around a shortcoming in GAE Standard? (yes, even though the
folder is `flexible`).

I'll work on removing the GCP docs for this.